requirements.txt
Flask==2.1.2
humanize==4.0.0
python_dateutil==2.8.2
requests==2.27.1
tenacity==8.0.1
timeloop_ng==1.0.0
waitress==2.1.2
apscheduler==3.9.1
